New server for multicoin API
