+++ coin-api/internal/rpc/client.go
